# Change the method of IntList equals from toString to ListEquality.equals. In my case, codegen 15x faster