I have a 6000-line proto file, and it takes one minute to generate Dart code using the Dart proto plugin. After changing to the ListEquality.equals method, it only takes 4.6 seconds. If I completely remove the commentBlock call and compile it into an AOT (Ahead-of-Time) artifact, it only takes 200ms.

> ❯ time protoc --dart_out=./dart/lib1 test.proto
> protoc --dart_out=./dart/lib1 test.proto  62.60s user 0.75s system 100% cpu 1:03.31 total
> ❯ time protoc --dart_out=./dart/lib1 --plugin=protobuf.dart/protoc_plugin/bin/protoc-gen-dart test.proto
> protoc --dart_out=./dart/lib1  test.proto  4.61s user 0.32s system 125% cpu 3.918 total