@osa1 2025-02-05
Thanks @qpalwo.

In the internal verison of this library we improve this even more by scanning the lists backwards, as it's more likely for two paths to share a prefix than a suffix.

But this version is still a big improvement compared to the current implementation.

Are you able to sign the CLA so that we can merge this?